+++ code/mc6840.h
@@ -48,7 +48,7 @@
 
 
 void mc6840_init (void);
-uint8_t millis(void);
+uint16_t millis(void);
 uint32_t uptime(void);
 uint16_t get_geiger_pulses (void);
 char* get_usiv_str (uint16_t cpmin, char * des);
